Refactor: Extract aspect ratio handling to `AspectRatioFormatter` class
- Moved aspect ratio-related methods (e.g., `FormatAspectRatio`, `GCD`, and `GetFormattedAspectRatio`) to the new `AspectRatioFormatter` class.
- Updated references across the project to use `AspectRatioFormatter` instead of `ImageTitleFormatter`.
- Removed redundant code from `ImageTitleFormatter`.
